refactor(stats): split ShareUsageCard to stay under max-lines lint rule
Extract helpers, SVG icons, and self-contained sub-components
(BackgroundGlows, CardFooter) into share-card-utils.tsx.
